added "main.pyw"
I found it useful to have a single startup file, in particular for a graphical user interface. Notice that the window mode limits console output and you can run the startup file (with all dependencies installed) run via a double-click.

For a module, the default entry point would be "__main__.py", hence the name "main.pyw".